+++ edgy/testing/factory/__init__.py
@@ -1,4 +1,5 @@
 from .base import ModelFactory
 from .fields import FactoryField
+from .subfactory import SubFactory
 
-__all__ = ["ModelFactory", "FactoryField"]
+__all__ = ["ModelFactory", "FactoryField", "SubFactory"]